app/requirements.txt
fastapi
uvicorn
pymupdf
pdfplumber
python-docx
python-multipart
pinecone
openai
tiktoken
aiohttp
python-dotenv
nltk
